SelfDrivingCar(Self-implementation)/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
_pytorch_select=0.1=cpu_0
blas=1.0=mkl
ca-certificates=2020.12.5=h033912b_0
certifi=2020.12.5=py38h50d1736_1
cffi=1.14.4=py38h2125817_0
cycler=0.10.0=py38_0
docutils=0.16=py38h50d1736_3
freetype=2.10.4=ha233b18_0
gettext=0.19.8.1=h7937167_1005
giflib=5.2.1=hbcb3906_2
glib=2.66.4=h22858aa_2
glib-tools=2.66.4=h22858aa_2
gst-plugins-base=1.14.5=hb4a159a_2
gstreamer=1.18.3=h29a8107_0
intel-openmp=2019.4=233
jpeg=9d=hbcb3906_0
kivy=2.0.0=py38h693e360_1
kiwisolver=1.3.0=py38h23ab428_0
lcms2=2.11=h92f6f08_0
libcxx=11.0.1=habf9029_0
libedit=3.1.20191231=h1de35cc_1
libffi=3.3=hb1e8313_2
libglib=2.66.4=h7424822_2
libiconv=1.16=haf1e3a3_0
libmklml=2019.0.5=0
libpng=1.6.37=ha441bb4_0
libtiff=4.1.0=hcb84e12_1
libwebp=1.0.2=hd3bf737_5
llvm-openmp=10.0.0=h28b9765_0
lz4-c=1.9.3=h23ab428_0
matplotlib=3.3.2=hecd8cb5_0
matplotlib-base=3.3.2=py38h181983e_0
mkl=2019.4=233
mkl-service=2.3.0=py38h9ed2024_0
mkl_fft=1.2.0=py38hc64f4ea_0
mkl_random=1.1.1=py38h959d312_0
ncurses=6.2=h0a44026_1
ninja=1.10.2=py38hf7b0b51_0
numpy=1.19.2=py38h456fd55_0
numpy-base=1.19.2=py38hcfb5961_0
olefile=0.46=py_0
openssl=1.1.1i=h35c211d_0
pcre=8.44=hb1e8313_0
pillow=8.1.0=py38h5270095_0
pip=20.3.3=py38hecd8cb5_0
pycparser=2.20=py_2
pygments=2.7.4=pyhd8ed1ab_0
pyparsing=2.4.7=pyhd3eb1b0_0
python=3.8.5=h26836e1_1
python-dateutil=2.8.1=pyhd3eb1b0_0
python_abi=3.8=1_cp38
pytorch=1.6.0=cpu_py38hd70000b_0
readline=8.1=h9ed2024_0
sdl2=2.0.12=h2e338ed_1
sdl2_image=2.0.5=hf6f9221_2
sdl2_mixer=2.0.4=h6de7cb9_0
sdl2_ttf=2.0.15=hc42bb53_0
setuptools=52.0.0=py38hecd8cb5_0
six=1.15.0=py38hecd8cb5_0
smpeg2=2.0.0=hfc679d8_1
sqlite=3.33.0=hffcf06c_0
tk=8.6.10=hb0a8c7a_0
tornado=6.1=py38h9ed2024_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.5=h41d2c2f_0
